Refactor e2e utility functions to a single separate package
