+++ test/utils/elasticsearch.go
@@ -1,16 +1,11 @@
-package e2e
+package utils
 
 import (
 	"context"
 	"fmt"
-	"os"
-	"os/exec"
 	"testing"
-	"time"
 
-	consolev1 "github.com/openshift/api/console/v1"
 	loggingv1 "github.com/openshift/elasticsearch-operator/pkg/apis/logging/v1"
-	"github.com/openshift/elasticsearch-operator/test/utils"
 
 	"github.com/operator-framework/operator-sdk/pkg/test"
 
@@ -23,41 +18,9 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
-const (
-	retryInterval        = time.Second * 3
-	timeout              = time.Second * 300
-	cleanupRetryInterval = time.Second * 3
-	cleanupTimeout       = time.Second * 30
-	elasticsearchCRName  = "elasticsearch"
-	kibanaCRName         = "kibana"
-	exampleSecretsPath   = "/tmp/example-secrets"
-)
+const elasticsearchCRName = "elasticsearch"
 
-func registerSchemes(t *testing.T) {
-	elasticsearchList := &loggingv1.ElasticsearchList{}
-	err := test.AddToFrameworkScheme(loggingv1.SchemeBuilder.AddToScheme, elasticsearchList)
-	if err != nil {
-		t.Fatalf("failed to add custom resource scheme to framework: %v", err)
-	}
-
-	kibanaList := &loggingv1.KibanaList{}
-	err = test.AddToFrameworkScheme(loggingv1.SchemeBuilder.AddToScheme, kibanaList)
-	if err != nil {
-		t.Fatalf("failed to add custom resource scheme to framework: %v", err)
-	}
-
-	consoleLinkList := &consolev1.ConsoleLinkList{}
-	err = test.AddToFrameworkScheme(consolev1.Install, consoleLinkList)
-	if err != nil {
-		t.Fatalf("failed to add custom resource scheme to framework: %v", err)
-	}
-}
-
-func elasticsearchNameFor(uuid string) string {
-	return fmt.Sprintf("%s-%s", elasticsearchCRName, uuid)
-}
-
-func createElasticsearchCR(t *testing.T, f *test.Framework, ctx *test.Context, esUUID, dataUUID string, replicas int) (*loggingv1.Elasticsearch, error) {
+func CreateElasticsearchCR(t *testing.T, f *test.Framework, ctx *test.Context, esUUID, dataUUID string, replicas int) (*loggingv1.Elasticsearch, error) {
 	namespace, err := ctx.GetWatchNamespace()
 	if err != nil {
 		return nil, fmt.Errorf("Could not get namespace: %v", err)
@@ -122,8 +85,8 @@ func createElasticsearchCR(t *testing.T, f *test.Framework, ctx *test.Context, e
 
 	cleanupOpts := &test.CleanupOptions{
 		TestContext:   ctx,
-		Timeout:       cleanupTimeout,
-		RetryInterval: cleanupRetryInterval,
+		Timeout:       DefaultCleanupTimeout,
+		RetryInterval: DefaultCleanupRetryInterval,
 	}
 
 	err = f.Client.Create(context.TODO(), cr, cleanupOpts)
@@ -134,8 +97,8 @@ func createElasticsearchCR(t *testing.T, f *test.Framework, ctx *test.Context, e
 	return cr, nil
 }
 
-func updateElasticsearchSpec(t *testing.T, f *test.Framework, desired *loggingv1.Elasticsearch) error {
-	return wait.Poll(retryInterval, timeout, func() (bool, error) {
+func UpdateElasticsearchSpec(t *testing.T, f *test.Framework, desired *loggingv1.Elasticsearch) error {
+	return wait.Poll(DefaultRetryInterval, DefaultTimeout, func() (bool, error) {
 		current := &loggingv1.Elasticsearch{}
 		key := client.ObjectKey{Name: desired.GetName(), Namespace: desired.GetNamespace()}
 
@@ -168,7 +131,7 @@ func updateElasticsearchSpec(t *testing.T, f *test.Framework, desired *loggingv1
 }
 
 // Create the secret that would be generated by CLO normally
-func createElasticsearchSecret(t *testing.T, f *test.Framework, ctx *test.Context, uuid string) error {
+func CreateElasticsearchSecret(t *testing.T, f *test.Framework, ctx *test.Context, uuid string) error {
 	t.Log("Creating required secret")
 	namespace, err := ctx.GetWatchNamespace()
 	if err != nil {
@@ -180,7 +143,7 @@ func createElasticsearchSecret(t *testing.T, f *test.Framework, ctx *test.Contex
 		return err
 	}
 
-	elasticsearchSecret := utils.Secret(
+	elasticsearchSecret := NewSecret(
 		elasticsearchNameFor(uuid),
 		namespace,
 		map[string][]byte{
@@ -198,8 +161,8 @@ func createElasticsearchSecret(t *testing.T, f *test.Framework, ctx *test.Contex
 
 	cleanupOpts := &test.CleanupOptions{
 		TestContext:   ctx,
-		Timeout:       cleanupTimeout,
-		RetryInterval: cleanupRetryInterval,
+		Timeout:       DefaultCleanupTimeout,
+		RetryInterval: DefaultCleanupRetryInterval,
 	}
 
 	err = f.Client.Create(context.TODO(), elasticsearchSecret, cleanupOpts)
@@ -210,7 +173,7 @@ func createElasticsearchSecret(t *testing.T, f *test.Framework, ctx *test.Contex
 	return nil
 }
 
-func updateElasticsearchSecret(t *testing.T, f *test.Framework, ctx *test.Context, uuid string) error {
+func UpdateElasticsearchSecret(t *testing.T, f *test.Framework, ctx *test.Context, uuid string) error {
 	namespace, err := ctx.GetWatchNamespace()
 	if err != nil {
 		return fmt.Errorf("Could not get namespace: %v", err)
@@ -245,111 +208,6 @@ func updateElasticsearchSecret(t *testing.T, f *test.Framework, ctx *test.Contex
 	return nil
 }
 
-func createKibanaCR(namespace string) *loggingv1.Kibana {
-	cpuValue, _ := resource.ParseQuantity("100m")
-	memValue, _ := resource.ParseQuantity("256Mi")
-
-	return &loggingv1.Kibana{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "kibana",
-			Namespace: namespace,
-		},
-		Spec: loggingv1.KibanaSpec{
-			ManagementState: loggingv1.ManagementStateManaged,
-			Replicas:        1,
-			Resources: &corev1.ResourceRequirements{
-				Limits: corev1.ResourceList{
-					corev1.ResourceMemory: memValue,
-				},
-				Requests: corev1.ResourceList{
-					corev1.ResourceCPU:    cpuValue,
-					corev1.ResourceMemory: memValue,
-				},
-			},
-		},
-	}
-}
-
-func createKibanaSecret(f *test.Framework, ctx *test.Context, esUUID string) error {
-	namespace, err := ctx.GetWatchNamespace()
-	if err != nil {
-		return fmt.Errorf("Could not get namespace: %v", err)
-	}
-
-	kibanaSecret := utils.Secret(
-		kibanaCRName,
-		namespace,
-		map[string][]byte{
-			"key":  getCertificateContents("system.logging.kibana.key", esUUID),
-			"cert": getCertificateContents("system.logging.kibana.crt", esUUID),
-			"ca":   getCertificateContents("ca.crt", esUUID),
-		},
-	)
-
-	cleanupOpts := &test.CleanupOptions{
-		TestContext:   ctx,
-		Timeout:       cleanupTimeout,
-		RetryInterval: cleanupRetryInterval,
-	}
-
-	err = f.Client.Create(context.TODO(), kibanaSecret, cleanupOpts)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func createKibanaProxySecret(f *test.Framework, ctx *test.Context, esUUID string) error {
-	namespace, err := ctx.GetWatchNamespace()
-	if err != nil {
-		return fmt.Errorf("Could not get namespace: %v", err)
-
-	}
-
-	kibanaProxySecret := utils.Secret(
-		fmt.Sprintf("%s-proxy", kibanaCRName),
-		namespace,
-		map[string][]byte{
-			"server-key":     getCertificateContents("kibana-internal.key", esUUID),
-			"server-cert":    getCertificateContents("kibana-internal.crt", esUUID),
-			"session-secret": []byte("TG85VUMyUHBqbWJ1eXo1R1FBOGZtYTNLTmZFWDBmbkY="),
-		},
-	)
-
-	cleanupOpts := &test.CleanupOptions{
-		TestContext:   ctx,
-		Timeout:       cleanupTimeout,
-		RetryInterval: cleanupRetryInterval,
-	}
-
-	err = f.Client.Create(context.TODO(), kibanaProxySecret, cleanupOpts)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func generateCertificates(t *testing.T, namespace, uuid string) error {
-	workDir := fmt.Sprintf("%s/%s", exampleSecretsPath, uuid)
-	storeName := elasticsearchNameFor(uuid)
-
-	err := os.MkdirAll(workDir, os.ModePerm)
-	if err != nil {
-		return fmt.Errorf("failed to create certificate tmp dir: %s", err)
-	}
-
-	cmd := exec.Command("./hack/cert_generation.sh", workDir, namespace, storeName)
-	out, err := cmd.Output()
-	if err != nil {
-		return fmt.Errorf("failed to generate certificate for %q: %v\n:%s", storeName, err, string(out))
-	}
-
-	return nil
-}
-
-func getCertificateContents(name, uuid string) []byte {
-	filename := fmt.Sprintf("%s/%s/%s", exampleSecretsPath, uuid, name)
-	return utils.GetFileContents(filename)
+func elasticsearchNameFor(uuid string) string {
+	return fmt.Sprintf("%s-%s", elasticsearchCRName, uuid)
 }
